requirements.txt
pandas
selenium
beautifulsoup4
langchain
transformers
peft
trl
datasets
sentence-transformers
torch
scikit-learn
gradio
python-dotenv
bitsandbytes
accelerate
rouge-score
bert-score
torchmetrics
huggingface-hub